pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "civ-sim"
version = "0.1.0"
description = "Persistent chunked 2D settlement simulator with evolving agents."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pillow>=10.0",
  "pygame-ce>=2.5.2",
  "PyYAML>=6.0",
  "torch>=2.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[project.scripts]
civ-sim = "civ_sim.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
